+++ tests/test_all.nim
@@ -21,4 +21,5 @@ import t_cli_args
 import t_library_api
 import t_e2e_integration
 import t_code_quality
+import t_memory_safety
 import test1